replication: introduce replicaset name
The replicaset name is carried with replicaset UUID wherever any
sanity validations are needed like whether 2 instances belong to
the same replicaset.

Part of #5029

@TarantoolBot document
Title: `box.cfg.replicaset_name` and `box.info.replicaset.name`
The new option `box.cfg.replicaset_name` allows to assign the
replicaset name to a human-readable text value to be displayed in
the new info key - `box.info.replicaset.name` - and to be
validated when the instances in the replicaset connect to each
other.

When set, it has to match in all instances of the entire
replicaset.

If a name wasn't set on replicaset bootstrap (was forgotten or the
replicaset is upgraded from a version < 3.0), it can be set /
changed / dropped, but only when `box.cfg.force_recovery == true`
in all instances of the replicaset. After the name is installed
and all the instances synced, the `force_recovery` can be set back
to `false`.

Note that such way of name set/change will work only after the
instance is up and running. I.e. to set a name first time one has
to start the instances without the name in the config, then set
`force_recovery` to true and the new name in the config, wait for
it to be applied, and then set `force_recovery` back to false.

The name can be <= 63 symbols long, can consist only of chars
['0'-'9'], '-' and 'a'-'z'. It must start with a letter. When
upper-case letters are used in `box.cfg`, they are automatically
converted to lower-case. The names are host- and DNS-friendly.