replication: introduce instance name
The instance name is carried with instance UUID everywhere in the
replication protocols. It is visible in all other instances via
_cluster and is displayed in monitoring.

Part of #5029

@TarantoolBot document
Title: `box.cfg.instance_name` and `box.info.name`
The new option `box.cfg.instance_name` allows to assign the
instance name to a human-readable text value to be displayed in
the new info key - `box.info.name`. Instances can see names of
their peers in `box.info.replication[id].name`.

When set, it has to be unique in the instance's replicaset.

If a name wasn't set on instance bootstrap (was forgotten or the
instance is upgraded from a version < 3.0), it can be set /
changed / dropped, but only when `box.cfg.force_recovery == true`.
After the name is updated the `force_recovery` can be set back to
`false`.

Note that such way of name set/change will work only after the
instance is up and running. I.e. to set names first time one has
to start the instances without the names in their configs, then
set `force_recovery` to true and the new names in the configs,
wait for them to be applied, and then set `force_recovery` back to
false everywhere.

The name can be <= 63 symbols long, can consist only of chars
['0'-'9'], '-' and 'a'-'z'. It must start with a letter. When
upper-case letters are used in `box.cfg`, they are automatically
converted to lower-case. The names are host- and DNS-friendly.